--- Websites/cv-page/src/App.tsx
@@ -1,24 +1,0 @@
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Header from "./components/Header";
-import HomePage from "./pages/HomePage";
-import PortfolioPage from "./pages/PortfolioPage";
-import GlobalStyles from "./styles/GlobalStyles";
-
-function App() {
-  return (
-    <>
-      <GlobalStyles />
-      <BrowserRouter>
-        <Header />
-        <Routes>
-          <Route path="/" element={<HomePage />} />
-          <Route path="/portfolio" element={<PortfolioPage />} />
-          <Route path="*" element={<div>Not Found</div>} />
-          <Route />
-        </Routes>
-      </BrowserRouter>
-    </>
-  );
-}
-
-export default App;
