--- Websites/cv-page/src/components/Contacts.tsx
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-const StyledContacts = styled.div`
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-function Contacts() {
-  return <StyledContacts>Contacts</StyledContacts>;
-}
-
-export default Contacts;
