--- Websites/cv-page/src/components/ErrorPage.tsx
@@ -1,5 +1,0 @@
-function ErrorPage() {
-  return <div>ErrorPage</div>;
-}
-
-export default ErrorPage;
