--- Websites/cv-page/src/pages/PortfolioPage.tsx
@@ -1,5 +1,0 @@
-function PortfolioPage() {
-  return <div>PortfolioPage</div>;
-}
-
-export default PortfolioPage;
