--- Websites/portfolio-page/babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  presets: ["next/babel"],
-  plugins: [["styled-components", { ssr: true, displayName: true }]],
-};
