+++ Websites/portfolio-page/src/pages/index.tsx
@@ -1,15 +1,14 @@
-import StartSection from "../components/sections/StartSection";
-import HeaderSection from "../components/sections/HeaderSection";
-import AboutSection from "../components/sections/AboutSection";
-import PortfolioSection from "../components/sections/PortfolioSection";
-import ContactSection from "../components/sections/ContactSection";
-import { Link, Element, Events, scrollSpy } from "react-scroll";
+import StartSection from "../sections/StartSection";
+import HeaderSection from "../sections/HeaderSection";
+import AboutSection from "../sections/AboutSection";
+import PortfolioSection from "../sections/PortfolioSection";
+import ContactSection from "../sections/ContactSection";
+import { Element } from "react-scroll";
 import { useInView } from "react-intersection-observer";
 import { useEffect, useState, useRef, use } from "react";
 
 export default function Home() {
   const [activeSection, setActiveSection] = useState<string>("home");
-  const [isScrolling, setIsScrolling] = useState(false);
   const [refStart, inViewStart] = useInView({ threshold: 0.5 });
   const [refAbout, inViewAbout] = useInView({ threshold: 0.5 });
   const [refPortfolio, inViewPortfolio] = useInView({ threshold: 0.5 });
@@ -17,15 +16,15 @@ export default function Home() {
 
   useEffect(() => {
     if (inViewStart) setActiveSection("home");
-    else if (inViewAbout) setActiveSection("about");
-    else if (inViewPortfolio) setActiveSection("portfolio");
-    else if (inViewContact) setActiveSection("contact");
+    if (inViewAbout) setActiveSection("about");
+    if (inViewPortfolio) setActiveSection("portfolio");
+    if (inViewContact) setActiveSection("contact");
   }, [inViewStart, inViewAbout, inViewPortfolio, inViewContact]);
 
   return (
     <>
       <HeaderSection activeSection={activeSection} />
-      <div className="mx-[9rem] pl-[8rem]">
+      <div className=" pl-[8rem] max-w-[100rem] mx-auto">
         <Element name="home">
           <div ref={refStart}>
             <StartSection />
